move peer inside cookedpeer
