+++ service/tbc/peermanager.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/btcsuite/btcd/wire"
 
-	"github.com/hemilabs/heminetwork/service/tbc/peer"
+	"github.com/hemilabs/heminetwork/service/tbc/cookedpeer/peer"
 )
 
 const (
