+++ service/tbc/peermanager_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/btcsuite/btcd/wire"
 
-	"github.com/hemilabs/heminetwork/service/tbc/peer"
+	"github.com/hemilabs/heminetwork/service/tbc/cookedpeer/peer"
 )
 
 func ping(ctx context.Context, t *testing.T, p *peer.Peer) error {
